OPT(dirty attempt): memoize is_with_annex value
I had spotted that identical for-each-ref was ran twice during a simple
datalad status in a regular git repository... investigation shown that
multple code paths from dataset.py lead to the same is_with_annex.
Each call was taking about 45 msec on my speedy laptop.  Oh well, not that
much, but trying to run  datalad rerun of a single commit, resulted in
over 20 of such calls.  That is already a second for a single rerun.

Proposed ATM solution is not full.  I am not sure yet what solution we
want to see though (memoization and resetting upon update/add_remote/etc),
or some more universal check (anything in git has changed).  But just wanted
to see what breaks if anything, and either there would be a notable impact
on benchmarks