MNT: Do a belated post-release changelog shuffle
test_version.test__version__() fails for me locally without this
update.

This situation has caused Travis to fail in the past, but it doesn't
seem to be triggering a failure now [0].  There was talk of removing
the need for this step, but nothing has come of it yet as far as I
know, so I'm not sure why this test isn't failing on Travis.

[0]: Here's the latest build with master:
     https://travis-ci.org/datalad/datalad/builds/641361504