DOC: Prefer "working tree" to "worktree" in --help text
In the past, Git used "worktree" and "working tree" somewhat
interchangeably, but that was before the existence of the `git
worktree` command.  Now "working tree" is preferred. [*]

It's not worth the churn to update all the internal spots where we say
"worktree" (or "work tree") to mean "working tree", but let's at least
use "working tree" in the user-facing command documentation.

[*]: https://lore.kernel.org/git/xmqq4lj9a7b0.fsf@gitster-ct.c.googlers.com/