BF: cmdline: Avoid unintentional capitalization in command help
Our help text has lots of spots where the first word in a line is
incorrectly capitalized when the line ends with a colon.  The code
that does this has a comment that says this is to capitalize the first
word of all "headings", but it's not clear what those headings are.

Without this code, the help output of all of our current commands only
differs in cases of incorrect capitalization, so it looks like this
code is no longer serving its original purpose while causing issues
elsewhere.  Drop it.

Re: https://github.com/datalad/datalad/pull/4091#discussion_r371312477