DOC: examples: Format multi-line command-line snippets consistently
At this point, download-url and run are the only commands with
command-line examples that are rendered across multiple lines.
download-url continues the line with "\", while the run example does
not and indents two extra spaces.

Either of these are probably okay, but we should be consistent.  Let's
go with the "\" form so that it's easier for users to paste examples
when following along locally.