DOC: examples: Consistently use unbound Python commands
The Python examples have a mix of dataset-bound method calls and
unbound calls (i.e., calls to datalad.api functions).  While the
dataset-bound calls are useful in general, using unbound calls in the
examples has the following benefits:

  * There are fewer unrepresented steps.

    The snippets aren't complete examples; they don't have setup code
    and the paths are placeholders.  With both unbound and bound
    methods, there are implied imports.  With the bound methods, there
    are additional missing pieces regarding what `ds` is, how the
    dataset was created, and what path it points to.

  * Unbound methods share the command-line path handling logic.

    A dataset string sent in through an unbound command should trigger
    the same relative path handling as the command line.  Using the
    unbound methods results in snippet pairs that are a closer match.
    The only assumption necessary is that the snippets are executed
    from the same current working directory.

On the other hand, a downside of _not_ using bound methods in the
examples is that users won't learn that bound methods exist through
the examples.  But snippets probably aren't a good means of
demonstrating this anyway, at least with their current capabilities.

Convert all the examples to unbound calls.  Note that some of the
converted status() examples add an explicit `dataset` argument either
because it's required for the intended behavior or because it's
helpful for comparison with neighboring examples.

Re: https://github.com/datalad/datalad/pull/4091#discussion_r371454318