RF: GitRepo.commit() based on WitlessRunner
Previous (positive) API and behavior are maintained, except for the lack
of 'protocol' support which is only available in the traditional runner.

In addition, when multiple call to git-commit need to be made, --amend
is now used from the second call onwards to avoid leaving multiple
commits behind for technical reasons, when conceptually there is only
as single changeset. Fixes gh-3914

This is a trimmed down changeset from a previous attempt to enable
interactive commit messages. Compare to
aa86a8d0df2af95399418f771f592aedbb848df1 to recover the discarded diff.