ENH: update: Use same merge approach for plain and annex repos
update(..., merge=True) merges in changes differently depending on
whether the repository is an annex repo.  For plain git repos, it
brings in the changes with 'git pull'.  For annex repos, it pulls with
'git annex sync', disabling committing and pushing.

This isn't problematic by itself, but it does mean that plain and
annex repos have a non-obvious discrepancy in behavior.  It also
prevents us from having finer control of what gets merged in for annex
repos, something that's needed in the upcoming commits.

Use 'git pull' for annex repos as well, except for repos that are on
adjusted branches.  Adjusted branches are tricky with submodules (as
touched on a bit in gh-3818) because an adjusted branch is, by design,
frequently rebased on top of its main branch, which means the
submodule ID that is being tracked is not stable.  There is likely no
good way to handle it, but merging an adjusted branch (or even its
main branch) ourselves will make things worse, so continue to use 'git
annex sync' [*].

[*] Another option would be to use 'git annex merge <revision>' for
    git-annex versions that support it (>= 7.20190819).  However, when
    testing that locally, I noticed that it failed with submodule
    conflicts in cases where 'git annex sync' did not.