ENH: update: Prefer merge of submodule revision to a ref without it
The last commit covers the case where we need an explicit fetch of a
revision that doesn't exist.  But when it exists, it's possible that
the revision is not contained in the merge target (e.g. it may be in
another branch), leading to the same problem described in the last
commit.

Check whether a branch merge target has the target revision.  If it
doesn't, change the merge base to the revision.

Note that this doesn't help on adjusted branches, where changes are
brought in with 'git annex sync <remote>'.