BF: update: Query intended branch.*.remote variable
When operating on a non-annex repo, update(..., merge=True) is
designed to check whether the configured remote for the current branch
is the same as the (given or inferred) sibling, running 'git pull
remote' without a refspec if it is.  However, branch.REMOTE.remote is
mistakenly queried rather than branch.BRANCH.remote.

Correct the variable name.  Note that an upcoming commit in this
series will get rid of this logic entirely, but it still seems worth
having a dedicated fix for this because that commit will be an
aggressive change that might not end up landing.