ENH: update: Avoid calling 'git pull' to merge changes
On update(..., merge=True) with a non-annex repo, we bring in changes
with (1) 'git pull <remote>' or (2) 'git pull <remote> <current
branch>', where <remote> is the resolved sibling.  We decide to use
the first if the current branch has a configured upstream remote that
matches the sibling.

This approach has the following issues:

  * it makes another remote connection even though we just fetched
    changes

  * it repeats inspection of some information that we already gathered
    (at least indirectly) in upstream code

Rework the git-pull-based merge functions to take a merge target and
call 'git merge' with the appropriate target to preserve the previous
behavior.