ENH: update: Try to fetch submodule commit explicitly if needed
update() calls either 'git fetch <resolved sibling>' or (a rough
approximation of) 'git fetch --all', depending on whether a sibling is
specified, what upstream is configured for the current branch, and
whether there is only a single remote.  In the case of a subdataset,
either form may fail to bring down the revision that's recorded in the
parent.  This can happen based on the fetch configuration or because
the revision is not within a branch on the remote.  The latter is an
easy state to get into due to Git's tendency to detach in submodules.

When a merge is requested, this leads to a confusing outcome: the
local parent repo replaces the submodule ID recorded in the remote
repo with the current HEAD of the submodule, a commit that doesn't
contain the revision recorded on the remote.  If the selected merge
target hasn't diverged from the local branch, the new local ID can
even match the one from before the update() call.  In other words, the
same update() call can bring in a new submodule ID to a parent, just
to revert it again when it updates the submodule.

Deal with this the same way that 'git submodule update' does.  If the
first fetch doesn't bring in the revision, try to fetch the revision
explicitly from the remote.  If that works and the merge flag is set,
use the revision as the merge target because none of the fetched refs
contain it.

When on an adjusted branch, don't "merge" the explicitly fetch
revision because we use git-annex sync (for reasons described a few
commits back), which can't be expected to bring the revision in.