BF: update: Don't consider "." as an upstream remote
branch.*.remote is "." when a branch has another local branch as its
upstream branch.  It doesn't make sense to consider the local
repository as a remote for an update, and it wouldn't work with the
downstream code anyway.  Make use of the new remote_only parameter of
get_tracking_branch() to ignore this case.