ENH: update: Try to fetch submodule commit explicitly if needed
update() calls either 'git fetch <resolved sibling>' or (a rough
approximation of) 'git fetch --all', depending on whether a sibling is
specified, what upstream is configured for the current branch, and
whether there is only a single remote.  In the case of a subdataset,
either form may fail to bring down the revision that's recorded in the
parent.  This can happen based on the fetch configuration or because
the revision is not within a branch on the remote.  The latter is an
easy state to get into due to Git's tendency to detach in submodules.

When a merge with --follow=parentds is requested, the update will fail
because the merge target (the revision) isn't available locally.  Deal
with this the same way that 'git submodule update' does: if the first
fetch doesn't bring in the revision, try to fetch the revision
explicitly from the remote.

When on an adjusted branch, don't "merge" the explicitly fetch
revision because we use git-annex sync (for reasons described a few
commits back), which can't be expected to bring the revision in.