ENH: update: Avoid calling 'git pull' to merge changes
On update(..., merge=True) with a non-annex repo, the _update_repo()
helper brings in changes with (1) 'git pull <remote>' or (2) 'git pull
<remote> <current branch>', where <remote> is the resolved sibling.
It decides to use the first if the current branch has a configured
upstream remote that matches the sibling.

This approach has the following issues:

  * it makes another remote connection even though we just fetched
    changes

  * it repeats inspection of some information that we already gathered
    (at least indirectly) in upstream code

Rework _update_repo() to take the merge target and call 'git merge'
with the appropriate target to preserve the previous behavior.