RF: WitlessRunner reimplemented with asyncio
Previous implementation had deadlock issues that I could not resolve.
This code is inspired by the highly informative

https://pymotw.com/3/asyncio/subprocesses.html

Reimplementations for the previous output processors are provided too.
They are now using "protocols" that are natively provided by asyncio,
and offer a conceptually similar, but extremely superior and more
flexible approach to the problem.

A no longer relevant test case was removed.