NF: update: Add option to merge in recorded submodule commit
`update --merge --recursive` tries to match the current branch with a
branch on the sibling and merge in those changes.  That is, it's
similar in spirit to

    git pull --no-recurse-submodules
    git submodule update --remote --merge --recursive

However, there are various situations where the selected merge target
does not point to the revision recorded in the parent (e.g., when the
sibling has switched to another branch or the scenario described at
gh-4148).

In these cases, if the caller wants to reflect the state of the remote
dataset hierarchy, the current behavior of update() won't do.  Add a
new option, --follow=parentds, that is more suited to the above
scenarios.  The new behavior is similar to

    git pull --no-recurse-submodules
    git submodule update --merge --recursive

The default value, --follow=sibling, is not intended to have any
change in behavior.

Rather than add a new option, we could extend --merge to take a value
that specifies the behavior (e.g., with --merge translating to
--merge=sibling).  But an upcoming commit will make a similar change
to add a "fast-forward only" option.  That's orthogonal to the
behavior being added in this commit, so use a separate option to avoid
needing to come up with a way for --merge to take multiple values
(e.g., "--merge=parentds,ff-only").

Note that --follow=parentds is marked as incompatible with adjusted
branches, where changes are brought in with 'git annex sync <remote>'
rather than `git merge <target>`.

Closes #4148.

Supersedes #4118 (a hybrid approach where the recorded subdataset
revision would be taken as the merge target if the selected sibling
branch did not contain the recorded revision)