ENH: assert_in_results: Make failure message more readable
assert_in_results() packs the repr() of the results into the message.
Given that the result records are often quite long, it can be hard to
parse and spot the relevant bits of information.  Make both the query
result and observed results easier to digest by using the same
json.dumps()-based formatting that assert_result_count() does.