TST: update: Make more specific assertions about result counts
Many spots in the update() tests assert that there are a certain
number of results of _any_ action type, sometimes with the "type" and
"status" specified.  This is fragile because it needs to be changed if
update() starts to yield other results (as it will in the next
commit).

Update these spots to specify that the action of interest is "update".