ENH: update: Yield result for merge event
When merging, update() doesn't provide any information to the callers
aside from an "Applying updates ..." log message.  And if a merge
fails, the call to update() aborts.  Instead let's provide a result
record for the merge event, and, for a failing merge, translate the
CommandError to a "status=error" result.  This will allow callers to
control it with `on_failure`.

Note that, if a subdataset sees a failure, it is not added to the
paths to save because the user will need to deal with the conflicts.
This will often lead to the top-level dataset also having a conflict,
in which case no save at all will happen.  But see the new
test_merge_conflict_in_subdataset_only for an example where the parent
dataset doesn't have a conflict despite its subdataset having one.