RF: Discontinue use of GitCommandError
Was only ever used to fake a GitPython error, but was nowhere acted
upon, specifically. It is now replaced with ValueError, as it is done
in other places when inappropriate arguments are given (here a non-empty
directory).

Also discontinue the practice of logging a warning about a situation
that will unconditionally raise an exception on the next line. If that
exception is caught anywhere, it is the responsibility of the caller to
decide if and how to log this event.