OPT: use lru_cache for config.get_git_version
I saw us invoking git version  over and over again.  There really should be
no reason to sample that "densly" and IMHO it should be safe to assume that
it would not drastically change through the life time of the process.
Moreover, we are using it seems only to check if git  >= 2.8, which is from
2016 .  I think it is most sensible to assume that we would have more recent
one, especially given that we need git annex >= 201905 ATM

Modified-by: Kyle Meyer <kyle@kyleam.com>
  Switched from custom decorator to lru_cache.