RF: Decomplexify configure_origins()
Previous code looked at two datasets simulatenously per iteration.
This could result in making a fetch attempt to a non-local repository,
because it would no check whether the second dataset was actually local.

The new code focuses on a single dataset at a time, and all checks are
performed on each one.