RF+OPT: Make WitlessProtocol respect datalad.log.outputs config setting
Done in a fashion to be consistent with old Runner behavior/formatting.
OPT: check log level (and now config) upon instantiation of WitlessProtocol,
and bind _log_nolog if logging of status is not expected.  This should make it
consume few less cycles in a tight loop of data coming in.