TST: Work around annex.dotfiles-related failures on adjusted branches
These tests are failing on the Windows GitHub CI build with
8.20200226-g2d3ef2c07.  Due to the combination of an adjusted branch
and us using a `-c annex.dotfiles=true` override in our calls to
git-annex, the repos end up in an unexpectedly dirty state when on an
adjusted branch.  (Dotfiles switch from pointer files back to
git-tracked files.)

The issue goes away if we set annex.dotfiles in .git/config or
git-annex's config.  So the two main options seem to be to (1) not set
annex.dotfiles=true or (2) to set it more persistently.  The former
isn't likely to be viable because it would prevent use from annexing
content like ".datalad/metadata/objects/", and the latter needs more
thought.  For now, make these tests pass by adjusting them to set
annex.dotfiles=true in .git/config when running on an adjusted branch.

Re: https://github.com/datalad/datalad/pull/4214#issuecomment-592149004