BF: publish: Don't fetch if remote.*.fetch is not set
To deal with a situation where fetching an empty repository (in
particular an git-lfs special annex remote) failed with

    fatal: Couldn't find remote ref HEAD" if no branch

the previous commit updated one of the fetch() calls in
_publish_dataset() to catch a CommandError and swallow it if it looks
like it was due to the above failure.

The core path to this error is fetching from an empty repository that
has neither remote.*.fetch nor an upstream configured.  Not having the
former is rare because 'clone' and 'remote add' both set
remote.*.fetch automatically.  Having remote.*.fetch unset but a
configured upstream would be a hard state to get into [*], but, in
that case, fetching doesn't serve any purpose for publish() because
nothing in refs/remotes/<remote> would be updated, only FETCH_HEAD.

Given publish() can't do anything useful without a fetch refspec,
let's instead avoid the above error by guarding the fetch() calls in
publish() with a check that a fetch refspec is set.

A minor though nice benefit of this approach is that it works with
both master (which raises a CommandError in the above case) and maint
(where the GitPython-based fetch raises an AssertionError when no
fetch refspec is configured).

Re: https://github.com/datalad/datalad/pull/4200#discussion_r384673198

[*] `git branch --set-upstream-to` would reject it, so it'd have to be
     set up directly via .git/config values.