TST: Relax get_open_files() test on Windows due to performance issues.
Rational in the diff. We would need to budget ~1min per call to
get_open_files. This isn't worth it, given that this is only an
occasionally used test helper.

The test is not cut short on windows, only performing the timing
insensitive parts. It still runs for almost 3min...

Fixes gh-4233