DOC: annexrepo: Tweak a debug message from ff1c6c99be
If there is no corresponding branch, this debug message is
mis-formatted as

  [DEBUG  ] Sync local 'git-annex' branch' branch.

We could avoid that by adding a third `if have_corresponding_branch`
to the final argument, but, at that point, it seems clearer to just
stitch together the corresponding branch-related arguments outside of
the debug format string.

(And, while not really a concern for one debug statement, rough timeit
measurements suggest an if-else and a .format() performs about the
same as repeating that if-else three times).