BF: annexrepo: Set annex.dotfiles=true in .git/config
As of version 8.20200226, git-annex no longer has --include-dotfiles
and, instead of skipping dotfiles that are not explicitly given as
paths, now adds dotfiles to git by default.  When the new
annex.dotfiles option is configured to true, git-annex decides the
fate of dotfiles based on the normal annex.largefiles mechanism.

97c28c7bc5 (RF: annexrepo: Adjust for removal of --include-dotfiles,
gh-4214) dealt with the removal of --include-dotfiles, but it only
_partially_ dealt with the new "dot files to git" behavior.  That
commit set annex.dotfiles=true for our calls to git-annex.  However,
when we set this transiently, later actions can result in annexed
dotfiles being added as regular git files [0], a problem that was
hinted at by the workarounds needed for test_run.test_sidecar and
test_save.test_save_dotfiles in adjusted mode.

We rely on annexing dotfiles (e.g., `.datalad/metadata/objects`), so
it seems we can't get around setting annex.dotfiles=true in the
repository.  When AnnexRepo is initialized, check whether
annex.dotfiles is configured in .git/config.  If it's not, set it to
true unless it has explicitly been set to false outside of DataLad.

Note that this solution doesn't consider the annex.dotfiles
configuration that could be stored in git-annex:config.log, which is
overridden by the value in .git/config.  This could be added to the
check if we decide the correctness is worth the expense for each
AnnexRepo initialization.

[0] Two posts on the git-annex site have discussion and demo scripts
    related to this:
    https://git-annex.branchable.com/forum/Get_annex.dotfiles__61__true_behavior_without_persistent_configuration__63__/
    https://git-annex.branchable.com/bugs/Guard_against_previously_annexed_dotfiles_being_converted_to_git_files__63__/