BF+RF: AnnexRepo - add set_shared_connection to (effective) set_remote_url
(effective) because apparently we had 2 set_remote_url definitions.
Earlier (in time and in the code) was located among other methods dealing
with remotes (thus I kept that location, which also invoked
self._set_shared_connection.

And later one, which added custom to AnnexRepo handling  of annexurl, but
lacked invocation of self._set_shared_connection.  I took later version,
placed it inplace of the older (earlier in the code) one, while maintaining
the self._set_shared_connection invocation.

This now causes ssh call to have socket defined for target2, so later
enableremote has a better chance to reuse it.

The hope is that it would close #4195 and resolve regression/difference
in behavior of the recent git-annex-standalone build in #4116