BF: annexrepo: Future-proof "has --include-dotfiles" check
AnnexRepo checks whether the git-annex version is at or below
7.20200226 to decide whether the installed git-annex has
--include-dotfiles.  That was a poorly chosen check because the
annex.dotfiles/--include-dotfiles change happened with the transition
to v8 (currently git-annex's master), which means DataLad will need a
compatibility fix if another release is cut for v7.  (This is
particuarly relevant given that 7.20200309 has been tagged in the
git-annex repo.)

Make this check use the v8 boundary.