TST: Restrict adjusted branch/annex.dotfiles workaround
The kludges from 547c1295cd (TST: Work around annex.dotfiles-related
failures on adjusted branches, 2020-02-28) are no longer necessary as
of git-annex's 1978a2420 (Fix bug that caused unlocked annexed
dotfiles to be added to git by the smudge filter when annex.dotfiles
was not set, 2020-03-09).  Avoid them if the annex version is newer
than the current release, 8.20200226.