TST: save: Skip two dotfiles checks with older git-annex versions
When this test runs on an adjusted branch and `to_git` is True or None
(i.e., the files should go to the annex), the repo is unexpectedly
dirty because the annexed files change to regular files in the working
tree.  The Windows run on GitHub CI, which currently has
7.20191107-g8ea269ef7, shows this failure [0], and I can also trigger
it locally with tools/eval_under_testloopfs.

This is resolved in the latest git-annex, so let's skip these failing
cases for older versions.

[0]: https://github.com/datalad/datalad/runs/495930993?check_suite_focus=true#step:8:665