ENH: Record state of a submodule's corresponding branch in the parent
Otherwise any subdataset in adjusted mode will have a state recorded
that belongs to a branch that is continuously rebased.

Also adjust DataLad's own diff-functionality to detect this intentional
'modified' state and report it as 'clean'.

This is a major step toward a first serious attempt to support nested
dataset in adjusted mode