RF: AnnexRepo.get_corresponding_branch() -> None without a managed branch
When dealing with managed branches, one frequently has to switch
behavior depending on their presence. With this change the return value
of get_corresponding_branch() carries the information whether there is
such a branch or not, and immediately allows decision making based on
it.

Previously, it required an additional call to is_managed_branch()
(that is duplicated in get_corresponding_branch() in order to determine
this condition.

This cost reduction is minimal, but the code complexity goes down.
The previous behavior is easily achieved via

    `get_corresponding_branch(branch) or branch`

and seem to communicate more direct what is intended.

A GitRepo.get_corresponding_branch() that always returns None has been
added to further simplify decision making and avoid type-based
switching.