RF: annexrepo: Update _save_add() to send git=True through GitRepo
When _save_add() is called with git=True, we call `git annex add` with
`-c annex.largefiles=nothing`.  If we switch to the recommended
--force-small in this case, a good number of tests on Windows fail [0]
because .gitattributes is unexpectedly modified, with LF line endings
converted to CRLF [1].  The reason for this isn't clear.  For those
tests, core.autocrlf is set to true, so even if a checkout switches
the working tree version to CRLF, Git should not consider it modified.
Despite forcing git-annex to use an adjusted branch and trying various
attribute/eol settings, I can't trigger the test failures locally.

These failures don't happen when we go through GitRepo._save_add()
[2], so let's do that.  Note that GitRepo._save_add(), like
GitRepo.add_(), sets an annex.largefiles=nothing override, so the
potential problem mentioned in the previous commit applies here as
well.

[0]: https://github.com/datalad/datalad/runs/480413705
[1]: The following run has debugging focused on one test:
     https://github.com/datalad/datalad/runs/501030112?check_suite_focus=true#step:8:17
[2]: https://github.com/datalad/datalad/runs/501282075