BF(TST,workaround): skip test which checks that git-annex creates ssh socket
Situation is not yet clear either it is purely our Runner (which did
reveal possible situation where it would hang, see
https://github.com/datalad/datalad/pull/4265) or there is something
really new worth considering in how git-annex process "behaves".

This commit is intended for 0.12.3 (maint branch).
When merging to master, this commit changes should first be RFed to
see if using GitWitlessRunner (instead of the old one as here)
mitigates the issue.