ENH: gitrepo: Use matched line as message for log_progress()
When the GitProgress protocol receives a line like

    Receiving objects:  32% (22269/69588), 7.98 MiB | 7.98 MiB/s

it extracts anything after (M/N) as the message.  This message is used
to detect when a particular action is done, but it's also used as the
"message" parameter to log_progress().

The latter is problematic when running non-interactively---where the
messages are shown with standard logging rather than progress
bars---because many lines have an empty message, leading to many lines
of uninformative output (as in this example [0]).

Prevent this by using the entire matched line as the message.  This
leads to the non-interactive logging closely matching what the caller
would see from the command line.

[0]: https://github.com/datalad-handbook/book/issues/390