ENH: log_progress: Support dropping non-interactive messages
log_progress() conveniently uses plain log message instead of a
progress bar when operating non-interactively, but there can be a
mismatch between what makes sense to show with a progress bar tick and
what makes sense to show as a plain log message.  For operations with
a lot of ticks, log_progress(lgr.info, ...)  interactively is simply
moving the progress bar, but non-interactively it is dumping a level
of information to the logs that isn't desirable for INFO level
messages.

Make it possible to tell log_progress() machinery to filter out a
message when operating non-interactively if logging isn't enabled for
some specified level.  To continue the example above, specifying
log_progress(lgr.info, ..., noninteractive_level=logging.DEBUG) leads
to a filter rejecting the INFO-level log message when the logger is
configured above logging.DEBUG.