ENH: gitrepo: Restrict non-interactive log_progress() output
As mentioned in the previous commit and gh-4293, the non-interactive
output for log_progress() calls in GitProgress can be very verbose for
INFO-level output.  Use the newly added noninteractive_level parameter
to prevent everything but the "Start *ing [...]" messages from being
displayed unless logging is enabled for at least the DEBUG level.