BF(TST): we must allow avoid mocking dealing with a lock file
In the test_HTTPDownloader_basic we wanted to simulate the situation where
writing into the target file is raising an IOError.  For that we mocked all
open calls and returned a fake file-like object which would raise the exception
upon write.   Now with locking, fasteners need to work with a lock file,
and our mocking was getting in a way causing fasteners operation to fail.
With this change we avoid mocking while opening .lck files, which allows
for fasteners to establish a lock as it wants, but then still fail as before
for a downloaded file.