ENH: annexrepo: Allow specifying protocol if using GitWitlessRunner
As of af082fa9d8 (2020-02-26), _run_annex_command() callers can
specify that GitWitlessRunner should be used instead of Runner.  In
this case, a protocol class is constructed based on the boolean values
of log_std{out,err}.

However, in order to do any custom handling of the output (as is done
in the next commit), the caller has to be able to specify a custom
protocol.  Add a `protocol` parameter to _run_annex_command() so that
they can do so.