DOC: common_cfg: Fix description of log level for log.outputs
When this option was initially added in 755d5f3ee2 (ENH: log.outputs
config setting to trigger logging of the stdout/stderr [...],
2016-10-07), it was mostly true: stdout was logged at DEBUG, but
stderr was logged at DEBUG _or higher_, depending on the value of
expect_stderr/expect_fail.  As of 5d903d7322 (RF: Switch Runner's log
levels from ERROR and DEBUG to 11 and 9 respectively [...],
2017-10-05), stdout is logged at level 9 and stderr at level 9 or 11.

Instead of trying to keep this up to date, instead give a floor
logging level that should be sufficient to see the output and is
unlikely to become stale.  Use 5 because this is the level that the
WitlessRunner logs at.

While here, update the title text to make it clearer that this is a
boolean thing.