NF: clone/install --reckless shared-{group|all|...}
Makes it possible to set up datasets for collaborative write access
on install. This is implemented as a reckless mode, because it bypasses
the standard annex write protection feature (see gh-1893).

- rewrote the `reckless` option docs
- added example to the clone docs
- add test of permission setup validity

This fixes gh-1146 although it doesn't implement a dedicated `--shared`
option. And it fixes gh-1893, by providing a dedicated, documented mode
that marks the effected datasets as untrusted.