NF: @collect_method_callstats to assist in finding what using a property
The immediate usecase target is Dataset.repo property, which is known
to not be particularly cheap and used all over the codebase.

And here is an example from decorating .repo with this decorator and
running a quick test:

    $> python -m nose -s -v -x datalad/core/distributed/tests/test_clone.py:test_autoenabled_remote_msg
    datalad.core.distributed.tests.test_clone.test_autoenabled_remote_msg ... ok
    ...
    OK
    The cost of property repo:
     Total: 5 calls from 1 objects with 2 contexts taking 0.53 sec
      140338799900560 distribution/dataset.py:config: 4 from 2 lines
      140338799900560 distribution/dataset.py:is_installed: 1 from 1 lines

so we had 5 calls for the same object in two different contexts.
And it immediately points to having 2 lines in config to access it -
most likely could be locally bound to make it at least a little
more efficient.