BF: interface: Drop spurious messages about suppressed results
As shown by the example in gh-4352, there are a couple of issues with
the message about suppressed results introduced in 8d4f2d0803
(gh-4337, 2020-03-25):

  * With a number of similar results at or below render_n_repetitions
    (which is 10), a message about a _negative_ number of suppressed
    results is shown.

  * Right above the limit, a message about zero suppressed results is
    shown.

Change the "display suppression message?" condition to check whether
there was a number of similar messages beyond the threshold rather
than simply checking that there was a chain of similar messages.

Fixes #4352.