[wip] ENH: create-sibling: Set HEAD of sibling to local repo's branch
When a caller creates a sibling with a branch other than master
checked out, the sibling's HEAD ref still points to refs/heads/master.
Depending on what is pushed to the repo, this may stay an unborn
branch.  On clone, there will be nothing to be checked out, so the
cloned repo with stay on an unborn master (the same state you'd be in
right after `git init`).

This is somewhat confusing to the cloner, but it gets even more
problematic because, on maint, `datalad update` crashes with a
confusing error message from git (relayed by git-annex).  On master,
`datalad update` fails with a more reasonable "could not determine
merge target" result, though the situation still probably wouldn't be
obvious to the caller.

When create-sibling, it seems like a reasonable assumption that the
caller wants the current branches in the dataset hierarchy to be the
HEAD ref in the created siblings.  So do that, avoiding the issue
above.