BF: gitrepo: Abort if trying to update a submodule on unborn branch
If update_submodule() is called with a submodule that has an unborn
branch checked out, the get_hexsha() call fails with a ValueError.  If
we were to catch that, and proceed with subbranch{,_hexsha} as None,
the underlying 'git submodule update' would fail in this situation.
So catch the get_hexsha() value error and raise an exception with a
more appropriate message.

Note that in this case there is probably another ref that could be
checked out, but later commits in this series will take care of that
in higher level code.